move registry to pkg
