+++ pkg/client/grpc/balancer/swr.go
@@ -20,7 +20,7 @@ import (
 	"sync"
 
 	"github.com/douyu/jupiter/pkg/core/constant"
-	"github.com/douyu/jupiter/pkg/core/registry"
+	"github.com/douyu/jupiter/pkg/registry"
 	"github.com/douyu/jupiter/pkg/server"
 	"github.com/smallnest/weighted"
 	"google.golang.org/grpc/attributes"
