+++ pkg/registry/init.go
