+++ pkg/registry/nop.go
