+++ pkg/registry/service.go
